Finaly figured out, kindof (works for one but not another), to use alibrary that is not isntalled into the Arduino IDE tool installation area, but nearby and in the overall SmartHome project area so it can also be checked into GIT here. - ByteSwap library works, is a single .h file with the byteswapping C preprocessor macros.
The uniNodeTarget.ino has significant changes. 
EEPROM is now used to store node ID info, number of nodes in this
microcontroller, etc.
Xbee send message seems broken. Receive message still works. Receive
uses the EEPROM data to know if message is for this node or not.
Pin Change Interrupts are in use for AC line Zero-cross detection as
well as manual pushbuttons for early development of the load driver
reactions to commands.
PWM is working at 60Hz for two independently controlled outputs. I
changed FAN output from D9 to D5, Light output remians on D6. This was a
convenience as D5 and D6 are both on the same timer counter, which is
the easier one to get running at 60Hz. (very slightly longer than 60Hz,
which is good as this way the Zero-Cross signal should come in before
counter rolls over) May later go for a longer PWM cycle if that works
better, usign the Zero-Cross signal to know when to brute-force reset
PWM to start, and use different Max count, step size etc. when making
adjustments to the PWM duty cycle. Will think more about this.

Minor update to documentation for Arduino Pins chart, to indicate what
"output pins" from the timer counters are the PWM pins. The FAN change
from Arduino pin D9 to D5 is no tyet incorporated in the doc, and needs
to be updated.