Added wxSmartHomeServer.depend and wxSmartomeServer.layout, but am not real whaerus ylt they are. Perhaps Code::Blocks config files for the project. I hadn' tnoticed them previously.
For the *App and *Main files, have added a working function call during idle event to check the file size of the SmartHome event Log file and if that has changed then refresh the GUI text area. NOTE that the refresh does not currently first clear the text area as expceted, but new text is added automatically on the fly, as the log file changes transparently from an unrelated text editor task, so changes to file are detected and responded to.