timerone lib fixed
I'm not sure what the timerone library authors were thinking, bu tI
couldn't get it to work for me. I changed to to fast PWM mode, removed a
divide-by-two, and it works much better.
Also adding wallcontrol file here, which was accidentally left out of
previous checkin. As a reminder, this improves the touchscreen
"debounce" and other code cleanup.